EN-865 Update attribute factory
