+++ engine/src/main/java/org/entando/entando/aps/system/services/entity/model/EntityAttributeValidationDto.java
@@ -21,13 +21,42 @@ import com.agiletec.aps.system.common.entity.model.attribute.util.IAttributeVali
 import com.agiletec.aps.system.common.entity.model.attribute.util.NumberAttributeValidationRules;
 import com.agiletec.aps.system.common.entity.model.attribute.util.OgnlValidationRule;
 import com.agiletec.aps.system.common.entity.model.attribute.util.TextAttributeValidationRules;
+import java.util.Date;
 
 /**
  * @author E.Santoboni
  */
 public class EntityAttributeValidationDto {
 
+    private Integer minLength;
+    private Integer maxLength;
+    private String regex;
+    private String rangeStartString;
+    private String rangeEndString;
+    private String rangeStartStringAttribute;
+    private String rangeEndStringAttribute;
+    private String equalString;
+    private String equalStringAttribute;
+
+    private Date rangeStartDate;
+    private Date rangeEndDate;
+    private String rangeStartDateAttribute;
+    private String rangeEndDateAttribute;
+    private Date equalDate;
+    private String equalDateAttribute;
+
+    private Integer rangeStartNumber;
+    private String rangeStartNumberAttribute;
+    private Integer rangeEndNumber;
+    private String rangeEndNumberAttribute;
+    private Integer equalNumber;
+    private String equalNumberAttribute;
+
     private EntityAttributeOgnlValidationDto ognlValidation;
+
+    public EntityAttributeValidationDto() {
+
+    }
 
     public EntityAttributeValidationDto(IAttributeValidationRules validationRules) {
         OgnlValidationRule ognlValidationRule = validationRules.getOgnlValidationRule();
@@ -41,10 +70,8 @@ public class EntityAttributeValidationDto {
         if (null == validationRules) {
             return;
         }
-        EntityAttributeValidationDto validationDto = null;
         if (attribute.isTextAttribute()) {
             TextAttributeValidationRules textValRule = (TextAttributeValidationRules) validationRules;
-            /*
             if (null != textValRule.getMaxLength() && textValRule.getMaxLength() > -1) {
                 this.setMaxLength(textValRule.getMaxLength());
             }
@@ -58,6 +85,7 @@ public class EntityAttributeValidationDto {
             this.setRangeEndStringAttribute(textValRule.getRangeEndAttribute());
             this.setRangeStartStringAttribute(textValRule.getRangeStartAttribute());
             this.setEqualStringAttribute(textValRule.getValueAttribute());
+            /*
             if (attribute instanceof EnumeratorAttribute) {
                 EnumeratorAttribute enumeratorAttribute = (EnumeratorAttribute) attribute;
                 this.setEnumeratorStaticItems(enumeratorAttribute.getStaticItems());
@@ -68,30 +96,194 @@ public class EntityAttributeValidationDto {
         }
         if (attribute instanceof DateAttribute) {
             DateAttributeValidationRules dateValRule = (DateAttributeValidationRules) validationRules;
-            /*
             this.setRangeEndDate((Date) dateValRule.getRangeEnd());
             this.setRangeStartDate((Date) dateValRule.getRangeStart());
             this.setEqualDate((Date) dateValRule.getValue());
             this.setRangeEndDateAttribute(dateValRule.getRangeEndAttribute());
             this.setRangeStartDateAttribute(dateValRule.getRangeStartAttribute());
             this.setEqualDateAttribute(dateValRule.getValueAttribute());
-             */
         }
         if (attribute instanceof NumberAttribute) {
             NumberAttributeValidationRules nulValRule = (NumberAttributeValidationRules) validationRules;
-            /*
             this.setRangeEndNumber((Integer) nulValRule.getRangeEnd());
             this.setRangeStartNumber((Integer) nulValRule.getRangeStart());
             this.setEqualNumber((Integer) nulValRule.getValue());
             this.setRangeEndNumberAttribute(nulValRule.getRangeEndAttribute());
             this.setRangeStartNumberAttribute(nulValRule.getRangeStartAttribute());
             this.setEqualNumberAttribute(nulValRule.getValueAttribute());
-             */
         }
         OgnlValidationRule ognlValidationRule = validationRules.getOgnlValidationRule();
         if (null != ognlValidationRule) {
             this.setOgnlValidation(new EntityAttributeOgnlValidationDto(ognlValidationRule));
         }
+    }
+
+    public Integer getMinLength() {
+        return minLength;
+    }
+
+    public void setMinLength(Integer minLength) {
+        this.minLength = minLength;
+    }
+
+    public Integer getMaxLength() {
+        return maxLength;
+    }
+
+    public void setMaxLength(Integer maxLength) {
+        this.maxLength = maxLength;
+    }
+
+    public String getRegex() {
+        return regex;
+    }
+
+    public void setRegex(String regex) {
+        this.regex = regex;
+    }
+
+    public String getRangeStartString() {
+        return rangeStartString;
+    }
+
+    public void setRangeStartString(String rangeStartString) {
+        this.rangeStartString = rangeStartString;
+    }
+
+    public String getRangeEndString() {
+        return rangeEndString;
+    }
+
+    public void setRangeEndString(String rangeEndString) {
+        this.rangeEndString = rangeEndString;
+    }
+
+    public String getRangeStartStringAttribute() {
+        return rangeStartStringAttribute;
+    }
+
+    public void setRangeStartStringAttribute(String rangeStartStringAttribute) {
+        this.rangeStartStringAttribute = rangeStartStringAttribute;
+    }
+
+    public String getRangeEndStringAttribute() {
+        return rangeEndStringAttribute;
+    }
+
+    public void setRangeEndStringAttribute(String rangeEndStringAttribute) {
+        this.rangeEndStringAttribute = rangeEndStringAttribute;
+    }
+
+    public String getEqualString() {
+        return equalString;
+    }
+
+    public void setEqualString(String equalString) {
+        this.equalString = equalString;
+    }
+
+    public String getEqualStringAttribute() {
+        return equalStringAttribute;
+    }
+
+    public void setEqualStringAttribute(String equalStringAttribute) {
+        this.equalStringAttribute = equalStringAttribute;
+    }
+
+    public Date getRangeStartDate() {
+        return rangeStartDate;
+    }
+
+    public void setRangeStartDate(Date rangeStartDate) {
+        this.rangeStartDate = rangeStartDate;
+    }
+
+    public Date getRangeEndDate() {
+        return rangeEndDate;
+    }
+
+    public void setRangeEndDate(Date rangeEndDate) {
+        this.rangeEndDate = rangeEndDate;
+    }
+
+    public String getRangeStartDateAttribute() {
+        return rangeStartDateAttribute;
+    }
+
+    public void setRangeStartDateAttribute(String rangeStartDateAttribute) {
+        this.rangeStartDateAttribute = rangeStartDateAttribute;
+    }
+
+    public String getRangeEndDateAttribute() {
+        return rangeEndDateAttribute;
+    }
+
+    public void setRangeEndDateAttribute(String rangeEndDateAttribute) {
+        this.rangeEndDateAttribute = rangeEndDateAttribute;
+    }
+
+    public Date getEqualDate() {
+        return equalDate;
+    }
+
+    public void setEqualDate(Date equalDate) {
+        this.equalDate = equalDate;
+    }
+
+    public String getEqualDateAttribute() {
+        return equalDateAttribute;
+    }
+
+    public void setEqualDateAttribute(String equalDateAttribute) {
+        this.equalDateAttribute = equalDateAttribute;
+    }
+
+    public Integer getRangeStartNumber() {
+        return rangeStartNumber;
+    }
+
+    public void setRangeStartNumber(Integer rangeStartNumber) {
+        this.rangeStartNumber = rangeStartNumber;
+    }
+
+    public String getRangeStartNumberAttribute() {
+        return rangeStartNumberAttribute;
+    }
+
+    public void setRangeStartNumberAttribute(String rangeStartNumberAttribute) {
+        this.rangeStartNumberAttribute = rangeStartNumberAttribute;
+    }
+
+    public Integer getRangeEndNumber() {
+        return rangeEndNumber;
+    }
+
+    public void setRangeEndNumber(Integer rangeEndNumber) {
+        this.rangeEndNumber = rangeEndNumber;
+    }
+
+    public String getRangeEndNumberAttribute() {
+        return rangeEndNumberAttribute;
+    }
+
+    public void setRangeEndNumberAttribute(String rangeEndNumberAttribute) {
+        this.rangeEndNumberAttribute = rangeEndNumberAttribute;
+    }
+
+    public Integer getEqualNumber() {
+        return equalNumber;
+    }
+
+    public void setEqualNumber(Integer equalNumber) {
+        this.equalNumber = equalNumber;
+    }
+
+    public String getEqualNumberAttribute() {
+        return equalNumberAttribute;
+    }
+
+    public void setEqualNumberAttribute(String equalNumberAttribute) {
+        this.equalNumberAttribute = equalNumberAttribute;
     }
 
     public EntityAttributeOgnlValidationDto getOgnlValidation() {
