+++ engine/src/main/java/org/entando/entando/web/entity/EntityManagerController.java
@@ -33,6 +33,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+/**
+ * @author E.Santoboni
+ */
 @RestController
 @RequestMapping(value = "/entityManagers")
 public class EntityManagerController {
